+++ .changeset/four-goats-breathe.md
@@ -1,6 +1,9 @@
 ---
 "io-ipatente-practices": minor
 "@io-ipatente/ui": minor
+"io-ipatente-licences": patch
+"io-ipatente-payments": patch
+"io-ipatente-vehicles": patch
 ---
 
 practices list and detail
